requirements.txt
numpy==1.18.4
opencv-contrib-python==4.2.0.34
Pillow==7.1.2
responder==2.0.5